+++ sonar-pmd-plugin/src/test/java/org/sonar/plugins/pmd/PmdTemplateTest.java
@@ -34,7 +34,7 @@ class PmdTemplateTest {
 
     @ParameterizedTest
     @ValueSource(strings = {
-            "1.2", "5", "6", "7", "8", "9", "1.9", "10", "1.10", "11", "1.11", "12", "13", "14", "15", "16", "17", "18"
+            "6", "7", "8", "9", "1.9", "10", "1.10", "11", "1.11", "12", "13", "14", "15", "16", "17", "18", "19", "19-preview","20", "20-preview"
     })
     void verifyCanHandleJavaLanguageVersion(String javaVersion) {
         final LanguageVersionHandler languageVersionHandler = PmdTemplate